+++ sdk/sdk/api_client/api/data_products_output_ports/ingest_output_port_contract.py
@@ -23,7 +23,7 @@ def _get_kwargs(
 
     _kwargs: dict[str, Any] = {
         "method": "post",
-        "url": "/api/v2/data_products/{data_product_id}/output_ports/{id}/data_contract".format(
+        "url": "/api/v2/data_products/{data_product_id}/output_ports/{id}/data_contract/".format(
             data_product_id=quote(str(data_product_id), safe=""),
             id=quote(str(id), safe=""),
         ),